+++ src/main/java/com/goggles/orderservice/application/service/OrderQueryService.java
@@ -8,6 +8,5 @@ import org.springframework.data.domain.Page;
 
 public interface OrderQueryService {
   OrderDetailResult getOrderDetails(UUID orderId, UUID userId);
-
   Page<OrderListResult> getOrders(OrderListQuery orderListQuery);
 }
