+++ src/main/java/com/goggles/orderservice/application/service/impl/OrderQueryServiceImpl.java
@@ -7,7 +7,6 @@ import com.goggles.orderservice.application.dto.result.OrderListResult;
 import com.goggles.orderservice.application.service.OrderQueryService;
 import com.goggles.orderservice.domain.entity.Order;
 import com.goggles.orderservice.domain.repository.OrderRepository;
-import com.goggles.orderservice.infrastructure.repository.custom.OrderCustomRepository;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -19,7 +18,6 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 public class OrderQueryServiceImpl implements OrderQueryService {
   private final OrderRepository orderRepository;
-  private final OrderCustomRepository orderCustomRepository;
 
   @Override
   public OrderDetailResult getOrderDetails(UUID orderId, UUID userId) {
@@ -33,7 +31,7 @@ public class OrderQueryServiceImpl implements OrderQueryService {
 
   @Override
   public Page<OrderListResult> getOrders(OrderListQuery orderListQuery) {
-    Page<Order> orderPage = orderCustomRepository.getOrderPage(orderListQuery);
+    Page<Order> orderPage = orderRepository.getOrderPage(orderListQuery);
     return orderPage.map(OrderListResult::from);
   }
 }
